Movi Door y Proyectile a un nuevo directorio Items, ya que Proyectile se va a usar tanto para balas del jugador como las del enemigo.
