+++ Assets/Scripts/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 501f0346f86291b469c6674215fb96ce
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
